Fix score search: expect float instead of integer
Search by score is broken after the update of moving to Wilson score. This commit fixes it.

The issue is, that QPixel expects an integer as score (to_i). Hence `score:>0.7` are interpreted as `score:>0`, which is obviously wrong. The fix should be just changing from to_i to to_f.