Prevent phantom answer in notification list.
<<Phantom answer in my notification list

My notification list includes the following entry:

New answer to your question 'How many votes are required for special
actions...'

However, there is no answer for that question.

I imagine that the author posted an answer and remove it, but as far
as I know I am able to see delete posts.>>

<<That... looks like a funny bug.

I started writing an answer, containing mostly the number "1", because
that would be an (not really helpful, but correct) answer to your
question.

To test, whether there was an answer limit, I submitted my one-letter
post. Correctly, I was rejected.

But apparently the code enters the notification, before the checks on
length/whatever are applied. As you can see here, we add the
notification in line 20, but save it (and run the validations) only in
line 23.>>

https://software.codidact.com/questions/278892