Add .env for docker-compose.yml file and add default values for local development
Makes use of the automatic .env variable auto-loading feature for docker-compose where values stored in a .env file in the same directory as the docker-compose.yml file will be automatically loaded to allow easier customization of docker-compose files. This will make it easier for devs to customize their local environment without keeping track of changes that shouldn't be tracked in git.