Update edit_profile.erb
Usually, there should be a better way to call these scripts. I had tried by putting them in users.js, application.js. Although, it wasn't working. So, finally it left for me. But, I would request someone to take a look at it. Since, we just code HTML and Ruby in the page.