Update application.scss
pagination's button had a default background color earlier. So, it was white in dark theme. `.pagination li .active` used for those buttons. And, warning notice's text-color was close to white. But, it's hard to see white color. So, I changed default to black for dark and light both. `i` was used for icon color.