Update comments JS; close #619
Nasty. That nice shiny new Popup class? It may be shiny but it has nasty
bugs lurking in it. We got a nice performance boost by caching popups
and retrieving them later when we need them again (like for subsequent
keypresses, which can be milliseconds apart, hence the performance
boost). However, while we were updating the items and location of the
popup, we were keeping the same callback function.

Given that the callbacks to this are usually generated in a closure and
reliant on variables in that closure, if we don't update the callback then
we also end up keeping the initial values of those variables... like
cursor position. This is the source of #619 - we were keeping the
initial value of cursorPos, which was always 4, so we only ever replaced
the first 3 characters of username.

Fix is simple: update the callback when retrieving a cached popup.
QPixel.Popup.getPopup has been updated to do this.