Add elasticsearchable concern
Elasticsearch can be enabled and disabled site-wide. To prevent errors
when it is disabled (trying to update a model in the ES index), I've
used a mock here to not have to deal with the internal logic of ES
assigned callbacks.
Advantage is that it can be switched on and off while the site is
running. Only problem with that is that the ES DB must be resynced if
any changes were made while it was disabled, otherwise some actions will
cause error 500 since the update cannot be applied in the ES database.