Report 204 on missing community-specific assets
Communities like codegolf have custom features implemented for them in
assets/community/codegolf.js and .css. All communities attempt to load
this information, which used to trigger error 404. With this change,
these requests will now be answered with 204 (ok, empty response).

This fixes the issue that the tests fail on the 404 errors (which normal
browsers would silently ignore).