Duplicate the favicon to the actual assets
With default configuration, codidact (in dev mode) does not find its
favicon because it is looking in the wrong spot. For the codidact
servers I suspect the public folder is served by NGINX directly, so the
issue does not occur. By adding the favicon into assets, normal pages
will load correctly without errors and the system tests can run.