Move check for signed in user from preferences function to preference
Previously the QPixel.preferences function returned null when there
was no signed in user. This caused an error as the QPixel.preference
function tried to access properties of null. Now the check for a
signed in user happens in QPixel.preference and this function returns
null when there is no signed in user. The three places that currently
use this return value all behave correctly when there is no signed in
user - the null does not cause incorrect behaviour.