Separate out shared post validations
PostValidations are defined now as the shared validations between posts
and suggested edits.

For checking the minimum body length, I switched it to checking the
markdown body as that is a much better representation of the length. In
my testing even a body with just "Hello" ended up being long enough
after all the html that was added.

I have altered the if statements on the validations for title and body
length as those can be nil for suggested edits. This is not a problem,
because other validations on Post already catch when they are empty.

Additionally, the title and body were only extensively checked if the
post type supported tags. I have now changed that such that title and
body are always required for posts.