Update edit_posts.html
Clarify that it's the goal of the post itself which matters, and not that of its author. The author's goals may change, or they may have created a post which turned out not to be aligned with the goals they had. Once a post has been committed to Codidact, the author's goals cease, and the network's goals with the post take effect.