Fix Chromium-specific mathjax overflow
Not sure why this works, but MathJAX includes an assistive MathML
container, and _that_ container appears to be the source of the
overflow:

> https://github.com/mathjax/MathJax/issues/2936#issuecomment-1261596396

The bug can also be fixed by overriding the positioning to not be
`relative`, but I don't know if that has unintended consequences, and
testing if it breaks positionally dependent mathjax is far more
difficult than testing if it anything renders weirdly as a result of the
max width statement.

Firefox renders identically with or without it.

There is also a small chance (based on comments) that Safari is
affected by the same bug, but I use Linux and can't spin up a macOS VM
to check (thanks, Apple).

Closes #802